test(e2e): mock invoke bridge for library flows in features spec
